# internal/api/v1/openapi3/swagger.json
{
    "components": {
        "schemas": {
...
            "fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_LoginResponse": {
                "properties": {
                    "code": {
                        "description": "业务码，见 internal/middleware/error.go 矩阵",
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.LoginResponse"
                    },
                    "message": {
                        "description": "人类可读消息",
                        "type": "string"
                    },
                    "request_id": {
                        "type": "string"
                    },
                    "timestamp": {
                        "description": "unix 毫秒",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "fxcore_internal_api_v1_dto.LoginRequest": {
                "properties": {
                    "email": {
                        "maxLength": 255,
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 128,
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ],
                "type": "object"
            },
...
            "fxcore_internal_api_v1_dto.RefreshRequest": {
                "properties": {
                    "refresh_token": {
                        "type": "string"
                    }
                },
                "required": [
                    "refresh_token"
                ],
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "AI 驱动的加密货币交易智能体平台 API（v1）。认证：access token 存 HttpOnly Cookie；写操作需 X-Signature/X-Timestamp/X-Nonce 签名头（HMAC-SHA256，±5min，nonce 防重放）。",
        "title": "FXcore API",
        "version": "1.0"
    },
...
    "paths": {
        "/auth/login": {
            "post": {
                "description": "校验账密，下发 access token（HttpOnly Cookie）与 refresh token / sign_secret",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.LoginRequest",
                                        "summary": "body",
                                        "description": "登录凭据"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "登录凭据",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_LoginResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1001 参数错误"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1005 登录限流（5 次/分钟/IP）"
                    }
                },
                "summary": "登录",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "description": "携带 refresh_token 时服务端吊销；始终清除 access Cookie",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.RefreshRequest",
                                        "summary": "body",
                                        "description": "refresh token（可选，携带则吊销）"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "refresh token（可选，携带则吊销）"
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-any"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "登出",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/forgot-password": {
            "post": {
                "description": "向注册邮箱发送密码重置链接；邮箱不存在时同样返回成功（防枚举）。SMTP 未配置时为 dev 模式：链接写日志并随响应 dev_link 返回",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ForgotPasswordRequest",
                                        "summary": "body",
                                        "description": "注册邮箱"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "注册邮箱",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_ForgotPasswordResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1001 参数错误"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1005 限流（5 次/分钟/IP）"
                    }
                },
                "summary": "忘记密码",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/reset-password": {
            "post": {
                "description": "携带邮件链接中的一次性令牌设置新密码；成功后令牌立即作废",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ResetPasswordRequest",
                                        "summary": "body",
                                        "description": "重置令牌 + 新密码（≥8 位）"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "重置令牌 + 新密码（≥8 位）",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_ResetPasswordResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1001 参数错误 / 令牌无效或已过期"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1005 限流（5 次/分钟/IP）"
                    }
                },
                "summary": "重置密码",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "轮换 refresh token：旧 token 作废并返回新 token + 新 access token + sign_secret",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.RefreshRequest",
                                        "summary": "body",
                                        "description": "旧 refresh token"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "旧 refresh token",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_RefreshResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1001 参数错误"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1101 refresh token 无效或已轮换"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1005 刷新限流"
                    }
                },
                "summary": "刷新令牌",
                "tags": [
                    "auth"
                ]
            }
        },
...
        "/models": {
            "get": {
                "description": "分页? 否——返回全部未删除模型，支持 fields 白名单过滤",
                "parameters": [
                    {
                        "description": "逗号分隔的字段白名单，如 id,name",
                        "in": "query",
                        "name": "fields",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_AIModelDTOList"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1002 未认证"
                    }
                },
                "summary": "模型列表",
                "tags": [
                    "models"
                ]
            },
            "post": {
                "description": "API Key 仅以 RSA-OAEP(SHA-256) 密文落库；需要 X-Signature/X-Timestamp/X-Nonce 签名头",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.CreateModelRequest",
                                        "summary": "body",
                                        "description": "模型配置"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "模型配置",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_AIModelDTO"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1001 参数错误 / 签名缺失或重放"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1002 未认证"
                    }
                },
                "summary": "创建模型",
                "tags": [
                    "models"
                ]
            }
        },
...
        "/models/{id}": {
            "delete": {
                "description": "软删除",
                "parameters": [
                    {
                        "description": "模型 ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-any"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1004 模型不存在"
                    }
                },
                "summary": "删除模型",
                "tags": [
                    "models"
                ]
            },
            "put": {
                "description": "Key 轮换：传新 api_key 时用当前公钥重新加密；custom 需要 base_url",
                "parameters": [
                    {
                        "description": "模型 ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.CreateModelRequest",
                                        "summary": "body",
                                        "description": "模型配置"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "模型配置",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_AIModelDTO"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1001 参数错误"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1004 模型不存在"
                    }
                },
                "summary": "更新模型",
                "tags": [
                    "models"
                ]
            }
        },
...
        "/positions": {
            "get": {
                "description": "当前未平仓持仓，支持 symbol 与 fields 过滤",
                "parameters": [
                    {
                        "description": "按交易对过滤，如 BTC-USDT",
                        "in": "query",
                        "name": "symbol",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "逗号分隔字段白名单",
                        "in": "query",
                        "name": "fields",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ApiResponse-fxcore_internal_api_v1_dto_PositionDTOList"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "持仓列表",
                "tags": [
                    "positions"
                ]
            }
        },
...
        "/ws/dashboard": {
            "get": {
                "description": "鉴权：query token 为 JWT access token；5s 推一次仪表盘快照，30s 心跳",
                "responses": {
                    "101": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "WebSocket 升级成功"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/fxcore_internal_api_v1_dto.ErrorResponse"
                                }
                            }
                        },
                        "description": "1002 token 无效"
                    }
                },
                "summary": "WebSocket 实时推送",
                "tags": [
                    "dashboard"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:8080/api/v1"
        }
    ]
}